+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -4,7 +4,10 @@ import lombok.Data;
 import lombok.RequiredArgsConstructor;
 import ru.yandex.practicum.filmorate.validation.AfterCustomDate;
 
-import javax.validation.constraints.*;
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Positive;
+import javax.validation.constraints.Size;
 import java.time.LocalDate;
 import java.util.HashSet;
 import java.util.Set;
@@ -12,7 +15,7 @@ import java.util.Set;
 @Data
 @RequiredArgsConstructor
 public class Film {
-    private int id;
+    private Long id;
 
     @NotNull(message = "Название не может быть пустым")
     @NotBlank(message = "Название не может быть пустым")
